landcover-analysis/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
earthengine-api==0.1.401
torch
torchaudio
torchvision
segmentation_models_pytorch
numpy
six
async-timeout==4.0.3
asyncpg==0.29.0
sqlalchemy==2.0.31
greenlet==3.0.3
cloudinary==1.36.0